use uint16_t in NetServer, DuelClient

+++ gframe/netserver.cpp
@@ -11,7 +11,7 @@ event* NetServer::broadcast_ev = 0;
 evconnlistener* NetServer::listener = 0;
 DuelMode* NetServer::duel_mode = 0;
 unsigned char NetServer::net_server_write[SIZE_NETWORK_BUFFER];
-unsigned short NetServer::last_sent = 0;
+size_t NetServer::last_sent = 0;
 
 bool NetServer::StartServer(unsigned short port) {
 	if(net_evbase)
